% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast.R
\name{nowcast}
\alias{nowcast}
\alias{nowcast.csfmt_reporting_triangle_v3}
\title{Nowcast a reporting triangle into an ensemble}
\usage{
nowcast(x, ...)

\method{nowcast}{csfmt_reporting_triangle_v3}(x, max_delay, n_sim = 1000, ...)
}
\arguments{
\item{x}{A `csfmt_reporting_triangle_v3`.}

\item{...}{Passed to methods.}

\item{max_delay}{Delay horizon in weeks.}

\item{n_sim}{Number of nowcast draws.}
}
\description{
Nowcast a reporting triangle into an ensemble
}
