% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming_grammar.R
\name{q_value}
\alias{q_value}
\title{Quantile label -> probability (inverse of [q_label])}
\usage{
q_value(label)
}
\arguments{
\item{label}{Character vector of quantile labels, e.g. "q02x5".}
}
\value{
Numeric vector of probabilities.
}
\description{
Quantile label -> probability (inverse of [q_label])
}
